Split utility functions into seperate files
