+++ frontend/src/utils/error.ts
@@ -1,0 +1,10 @@
+import type { ApiError } from "../client"
+
+export const handleError = (err: ApiError, showToast: any) => {
+  const errDetail = (err.body as any)?.detail
+  let errorMessage = errDetail || "Something went wrong."
+  if (Array.isArray(errDetail) && errDetail.length > 0) {
+    errorMessage = errDetail[0].msg
+  }
+  showToast("Error", errorMessage, "error")
+}
